pyproject.toml
[project]
name = "mcp-server-demo"
version = "0.1.0"
description = "Local MCP server demo with weather and SQLite tools"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "mcp[cli]>=1.9.0",
  "httpx>=0.27.0",
  "python-dotenv>=1.0.1",
  "openai>=1.75.0",
  "streamlit>=1.45.0"
]

[project.scripts]
mcp-server-demo = "server:main"

[tool.setuptools]
py-modules = ["server", "client_openai_api", "web_client"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"